+++ java/src/main/java/edu/richmond/nlp/Writer/CSVDocumentDocumentWriter.java
@@ -16,7 +16,7 @@ public class CSVDocumentDocumentWriter {
 
   public CSVDocumentDocumentWriter(String docID) {
     this.docID = docID;
-    this.header = String.format("doc_id,time,version,language,uri%n");
+    this.header = String.format("id,time,version,language,uri%n");
   }
 
   public String print(Annotation doc) {